requirements.txt
Flask==0.12.2
flask-socketio
Flask-PyMongo
blinker
github-webhook==1.0.2
PyGithub==1.35
gunicorn==19.7.1
requests
